serving: Use chunked-encoding for progressive content under HTTP/1.1
For HTTP/1.0, the only way to generate content progressively is to rely on
the closing of connection to convey end of response. This can typically
lead to silently truncated content in HTTP client implementations.

For HTTP/1.1, the better way is to use `Transfer-Encoding: chunked` which
has a builtin mechanism to always terminate response with a 0-sized chunk,
marking end of stream.